requirements.txt
python-dotenv
langchain
langchain-google-genai
fastapi[all]